+++ examples/circle.rs
@@ -136,7 +136,7 @@ fn setup(
             sample_mapping: ImageSampleMapping::ModulateRGB,
         })
         .render(FlipbookModifier { sprite_grid_size })
-        .render(ColorOverLifetimeModifier { gradient })
+        .render(ColorOverLifetimeModifier::new(gradient))
         .render(SizeOverLifetimeModifier {
             gradient: Gradient::constant([0.5; 3].into()),
             screen_space_size: false,