+++ examples/spawn.rs
@@ -109,9 +109,7 @@ fn setup(
             .init(init_age1)
             .init(init_lifetime1)
             .update(update_accel1)
-            .render(ColorOverLifetimeModifier {
-                gradient: color_gradient1,
-            })
+            .render(ColorOverLifetimeModifier::new(color_gradient1))
             .render(SizeOverLifetimeModifier {
                 gradient: size_gradient1,
                 screen_space_size: false,
@@ -163,9 +161,7 @@ fn setup(
         .init(init_vel2)
         .init(init_age2)
         .init(init_lifetime2)
-        .render(ColorOverLifetimeModifier {
-            gradient: gradient2,
-        }),
+        .render(ColorOverLifetimeModifier::new(gradient2)),
     );
 
     commands
@@ -231,9 +227,7 @@ fn setup(
         .init(init_lifetime3)
         .init(init_size3)
         .update(update_accel3)
-        .render(ColorOverLifetimeModifier {
-            gradient: gradient3,
-        }),
+        .render(ColorOverLifetimeModifier::new(gradient3)),
     );
 
     commands